backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
PyJWT>=2.8.0
telethon==1.42.0
python-dotenv>=1.0.0
